Update IDX_ROUTINE_XXX and IDX_FIELD_XXX
Conventions for MAX_IDX_XXX have changed to be inclusive.  Also,
the details array is indexed started at 1 now.

(Note: the prefix IDX_ is associated with 1-based indexing, and
fields don't use that at present.  It might change if fields are
going to actually be OBJECT! at which point they would be 1-based.
Review later.)