Changes so simple GTK demo (mostly) works
Notably, Canon_Symbol() does not work on EXT_SYM_XXX IDs.  But the
case where it was trying to be applied actually had the symbol
already available, so it wasn't an issue.

This has problems with subcell addressing in terms of setting
elements of arrays.  This is likely going to be addressed by using
VECTOR! instead of BLOCK! for the arrays.